[feature] Added list of my nfts (MyNftCellsl)
