+++ iOS-FakeNFT-Extended/UI/Profile/ProfileMain/Views/ProfileViewMock.swift
@@ -29,7 +29,7 @@ struct ProfileViewMock {
         return ProfileView()
     }()
     
-    let mockNFT = NftItem(
+    static let mockNFT = NftItem(
         id: "594aaf01-5962-4ab7-a6b5-470ea37beb93",
         name: "Minnie Sanders",
         images: [
@@ -43,4 +43,43 @@ struct ProfileViewMock {
         createdAt: "2023-07-11T00:08:48.728Z[GMT]",
         description: "mediocritatem interdum eleifend penatibus adipiscing mattis"
     )
+    
+    static let mockNFTs: [NftItem] = [
+            NftItem(
+                id: "1",
+                name: "Lilo #1",
+                images: [
+                    "https://code.s3.yandex.net/Mobile/iOS/NFT/Pink/Lilo/1.png"
+                ],
+                rating: 2,
+                price: 40.59,
+                author: "Автор 1",
+                createdAt: nil,
+                description: nil
+            ),
+            NftItem(
+                id: "2",
+                name: "Ailo #2",
+                images: [
+                    "https://code.s3.yandex.net/Mobile/iOS/NFT/Pink/Lilo/2.png"
+                ],
+                rating: 4,
+                price: 55.00,
+                author: "Автор 2",
+                createdAt: nil,
+                description: nil
+            ),
+            NftItem(
+                id: "3",
+                name: "Filo #3",
+                images: [
+                    "https://code.s3.yandex.net/Mobile/iOS/NFT/Pink/Lilo/3.png"
+                ],
+                rating: 5,
+                price: 70.25,
+                author: "Автор 3",
+                createdAt: nil,
+                description: nil
+            )
+        ]
 }
